requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit-ace
networkx
pyvis
python-Levenshtein